+++ test/serializers/associations_test.rb
@@ -73,6 +73,15 @@ module ActiveModel
         end
       end
 
+      def test_association_with_namespace_options_uses_namespace_serializer
+        @post.comments = [BasicComment.new, SpecialComment.new]
+        @post_serializer = Test::Serializer::Post.new(@post)
+        @post_serializer.each_association do |name, serializer, options|
+          assert_kind_of Test::Serializer::BasicComment, serializer.to_a.first
+          assert_kind_of Test::Serializer::SpecialComment, serializer.to_a.last
+        end
+      end
+
       def test_belongs_to
         assert_equal(
           { post: { type: :belongs_to, association_options: {} },